refactor(map): extract shared subject-list pipeline into useMapSubjectList
The spiderfy-resolution → ranking → list + per-id ranking lookup was
duplicated verbatim between MapPageView and the meeting map page. Pull it
into a useMapSubjectList(allSubjects, visibleSubjects, spiderfiedIds) hook;
each surface keeps only its own visibleSubjects rule (viewport intersection
on /map vs. always-listed for the one meeting's municipality) and consumes
the identical downstream shaping from one place.
